+++ ASI.MGC.FS/Controllers/InvoiceController.cs
@@ -57,8 +57,8 @@ namespace ASI.MGC.FS.Controllers
                 objInvoiceMaster.INVNO_IPM = invNumber;
                 objInvoiceMaster.INVDATE_IPM = Convert.ToDateTime(DateTime.Now.ToShortDateString());
                 objInvoiceMaster.CUSTNAME_IPM = Convert.ToString(frm["CustDetail"]);
-                objInvoiceMaster.SHIPPING_IPM = Convert.ToInt32(frm["TotalShipCharges"]);
-                objInvoiceMaster.DISCOUNT_IPM = Convert.ToInt32(frm["TotalDiscount"]);
+                objInvoiceMaster.SHIPPING_IPM = Convert.ToDecimal(frm["TotalShipCharges"]);
+                objInvoiceMaster.DISCOUNT_IPM = Convert.ToDecimal(frm["TotalDiscount"]);
                 objInvoiceMaster.INVTYPE_IPM = "INV";
                 objInvoiceMaster.LPONO_IPM = _unitOfWork.Repository<MATERIALRECEIPTMASTER>().FindByID(mrvNumber).NOTES_MRV;
                 _unitOfWork.Repository<INVMASTER>().Insert(objInvoiceMaster);
